+++ tests/test_person_monitoring.py
@@ -1,18 +1,23 @@
 """
 Test for call monitoring settings
 """
+
 import asyncio
 import base64
+import json
 import random
 from concurrent.futures import ThreadPoolExecutor
 from contextlib import contextmanager
 from dataclasses import dataclass
 from typing import ClassVar, Optional
 
+from pydantic import TypeAdapter
+
 from tests.base import TestCaseWithUsers
 from wxc_sdk import WebexSimpleApi
 from wxc_sdk.all_types import *
 from wxc_sdk.as_api import AsWebexSimpleApi
+from wxc_sdk.person_settings.monitoring import MonitoringMember
 
 
 def print_monitoring(*, user: Person, monitoring: Monitoring):
@@ -21,26 +26,88 @@ def print_monitoring(*, user: Person, monitoring: Monitoring):
     for me in monitoring.monitored_elements or []:
         if me.member:
             print(f'    {me.member.display_name} ({me.member.member_type}) in {me.member.location_name}')
+        elif me.cpe:
+            print(f'    {me.cpe.name} (CPE) in {me.cpe.location}')
+        elif me.speed_dial:
+            print(f'    {me.speed_dial.display_name} (SD): {me.speed_dial}')
+            ...
         else:
-            print(f'    {me.cpe.name} (CPE) in {me.cpe.location_name}')
+            print(f'    unknown monitored element: {me}')
+    print('Data for update:')
+    print(json.dumps(monitoring.update(), indent=2))
 
 
 class TestRead(TestCaseWithUsers):
-
     def test_001_read_all(self):
         """
         Read monitoring setting of all users
         """
-        ps = self.api.person_settings
+        mo = self.api.person_settings.monitoring
 
         with ThreadPoolExecutor() as pool:
-            monitoring_settings = list(pool.map(lambda user: ps.monitoring.read(entity_id=user.person_id),
-                                                self.users))
+            monitoring_settings = list(pool.map(lambda user: mo.read(entity_id=user.person_id), self.users))
         print(f'Got monitoring details for {len(self.users)} users')
-        with_elements = [(user, ms) for user, ms in zip(self.users, monitoring_settings)
-                         if ms.monitored_elements]
+        with_elements = [
+            (user, ms) for user, ms in zip(self.users, monitoring_settings, strict=True) if ms.monitored_elements
+        ]
         for user, ms in with_elements:
             print_monitoring(user=user, monitoring=ms)
+            print()
+            if ms.monitored_elements:
+                # try to update
+                mo.configure(entity_id=user.person_id, settings=ms)
+                ms_after = mo.read(entity_id=user.person_id)
+                self.assertEqual(ms, ms_after)
+
+    def test_002_get_available_members_for_monitoring_all(self):
+        """
+        Get available members for monitoring for all users
+        """
+        mo = self.api.person_settings.monitoring
+        with ThreadPoolExecutor() as pool:
+            available_members = list(
+                pool.map(
+                    lambda user: list(mo.get_available_members_for_monitoring(entity_id=user.person_id)), self.users
+                )
+            )
+        print(f'Got available members for monitoring for {len(self.users)} users')
+        for user, members in zip(self.users, available_members, strict=True):
+            members: list[MonitoringMember]
+            print(f'**** user "{user.display_name}({user.emails[0]})"')
+            print(
+                json.dumps(
+                    TypeAdapter(list[MonitoringMember]).dump_python(
+                        members, mode='json', by_alias=True, exclude_none=True
+                    ),
+                    indent=2,
+                )
+            )
+            print()
+
+    def test_003_get_available_speed_dials_for_monitoring_all(self):
+        """
+        Get available speed dials for monitoring for all users
+        """
+        mo = self.api.person_settings.monitoring
+        with ThreadPoolExecutor() as pool:
+            available_members = list(
+                pool.map(
+                    lambda user: list(mo.get_available_speed_dials_for_monitoring(entity_id=user.person_id)), self.users
+                )
+            )
+        print(f'Got available speed dials for monitoring for {len(self.users)} users')
+        for user, members in zip(self.users, available_members, strict=True):
+            members: list[MonitoringMember]
+            print(f'**** user "{user.display_name}({user.emails[0]})"')
+            print(
+                json.dumps(
+                    TypeAdapter(list[MonitoringMember]).dump_python(
+                        members, mode='json', by_alias=True, exclude_none=True
+                    ),
+                    indent=2,
+                )
+            )
+            print()
 
 
 @dataclass(init=False, repr=False)
@@ -54,8 +121,10 @@ class TempCPE:
 
     async def cleanup(self):
         async with AsWebexSimpleApi(tokens=self.api.access_token) as api:
-            tasks = [api.telephony.callpark_extension.delete(location_id=self.location_id, cpe_id=cpe_id)
-                     for cpe_id in self.generated_cpe_ids]
+            tasks = [
+                api.telephony.callpark_extension.delete(location_id=self.location_id, cpe_id=cpe_id)
+                for cpe_id in self.generated_cpe_ids
+            ]
             self.generated_cpe_ids = []
             await asyncio.gather(*tasks)
 
@@ -74,9 +143,7 @@ class TempCPE:
                 ids_cpes_present = set(cpe.cpe_id for cpe in cpes_present)
 
                 # ids of CPEs which can be added: all cpes not currently used
-                available_ci_cpe_ids = [cpe.cpe_id
-                                        for cpe in existing_cpes
-                                        if cpe.cpe_id not in ids_cpes_present]
+                available_ci_cpe_ids = [cpe.cpe_id for cpe in existing_cpes if cpe.cpe_id not in ids_cpes_present]
 
                 # are we missing any to be able to fulfill the request?
                 missing = max(cpe_count - len(available_ci_cpe_ids), 0)
@@ -87,27 +154,25 @@ class TempCPE:
 
                     location_id = random.choice(locations).location_id
                     self.location_id = location_id
-                    cpe_names = set(cpe.name for cpe in existing_cpes
-                                    if cpe.location_id == location_id)
-                    new_names = (name for i in range(1000)
-                                 if (name := f'cpe_{i:03}') not in cpe_names)
+                    cpe_names = set(cpe.name for cpe in existing_cpes if cpe.location_id == location_id)
+                    new_names = (name for i in range(1000) if (name := f'cpe_{i:03}') not in cpe_names)
 
                     # assumption: we can assign new extensions to CPEs in ascending order
-                    extensions = [int(cpe.extension)
-                                  for cpe in existing_cpes
-                                  if cpe.location_id == location_id]
+                    extensions = [int(cpe.extension) for cpe in existing_cpes if cpe.location_id == location_id]
                     extensions = extensions or [1100]
                     max_cpe_extension = max(extensions)
 
-                    tasks = [api.telephony.callpark_extension.create(location_id=location_id,
-                                                                     name=next(new_names),
-                                                                     extension=str(max_cpe_extension + i + 1))
-                             for i in range(missing)]
+                    tasks = [
+                        api.telephony.callpark_extension.create(
+                            location_id=location_id, name=next(new_names), extension=str(max_cpe_extension + i + 1)
+                        )
+                        for i in range(missing)
+                    ]
                     self.generated_cpe_ids = await asyncio.gather(*tasks)
                     available_ci_cpe_ids.extend(self.generated_cpe_ids)
 
             result = random.sample(available_ci_cpe_ids, cpe_count)
-            return result
+            return result  # type: ignore[return-value]
 
         self.generated_cpe_ids = []
         cpe_list = asyncio.run(get_cpes())
@@ -178,10 +243,8 @@ class TestUpdate(TestCaseWithUsers):
             temp_cpe = TempCPE(api=self.api)
             with temp_cpe.generate_cpes(cpes_present=before.monitored_cpes, cpe_count=8) as new_cpe_ids:
                 # ths is what we want to add
-                new_monitoring_elements = [cpe_id
-                                           for cpe_id in new_cpe_ids]
-                settings = Monitoring(
-                    monitored_elements=(before.monitored_elements or []) + new_monitoring_elements)
+                new_monitoring_elements = [cpe_id for cpe_id in new_cpe_ids]
+                settings = Monitoring(monitored_elements=(before.monitored_elements or []) + new_monitoring_elements)
 
                 # update
                 mon.configure(entity_id=user.person_id, settings=settings)
@@ -212,17 +275,16 @@ class TestUpdate(TestCaseWithUsers):
             # get current settings
             before = mon.read(entity_id=user.person_id)
             present_ids = [m.member_id for m in before.monitored_members]
-            user_candidates = [uc for uc in self.users
-                               if uc.person_id not in present_ids and uc.person_id != user.person_id]
+            user_candidates = [
+                uc for uc in self.users if uc.person_id not in present_ids and uc.person_id != user.person_id
+            ]
             to_add = random.sample(user_candidates, 3)
             to_add: list[Person]
             print(f'Trying to add monitoring for: {", ".join(u.display_name for u in to_add)}')
 
             # ths is what we want to add
-            new_monitoring_elements = [user.person_id
-                                       for user in to_add]
-            settings = Monitoring(
-                monitored_elements=(before.monitored_elements or []) + new_monitoring_elements)
+            new_monitoring_elements = [user.person_id for user in to_add]
+            settings = Monitoring(monitored_elements=(before.monitored_elements or []) + new_monitoring_elements)
 
             # update
             mon.configure(entity_id=user.person_id, settings=settings)
@@ -250,16 +312,13 @@ class TestUpdate(TestCaseWithUsers):
             # get current settings
             before = mon.read(entity_id=user.person_id)
             present_ids = [m.member_id for m in before.monitored_members]
-            user_candidates = [user for user in self.users
-                               if user.person_id not in present_ids]
+            user_candidates = [user for user in self.users if user.person_id not in present_ids]
             to_add = random.sample(user_candidates, 3)
             print(f'Trying to add monitoring for: {", ".join(u.display_name for u in to_add)}')
 
             # ths is what we want to add
-            new_monitoring_elements = [user.person_id
-                                       for user in to_add]
-            settings = Monitoring(
-                monitored_elements=(before.monitored_elements or []) + new_monitoring_elements)
+            new_monitoring_elements = [user.person_id for user in to_add]
+            settings = Monitoring(monitored_elements=(before.monitored_elements or []) + new_monitoring_elements)
 
             # update
             mon.configure(entity_id=user.person_id, settings=settings)
@@ -268,8 +327,9 @@ class TestUpdate(TestCaseWithUsers):
             after = mon.read(entity_id=user.person_id)
             print_monitoring(user=user, monitoring=after)
 
-        decoded_user_ids = list(map(lambda member: base64.b64decode(member.member_id + '==').decode(),
-                                    after.monitored_members))
-        for member, decoded in zip(after.monitored_members, decoded_user_ids):
+        decoded_user_ids = list(
+            map(lambda member: base64.b64decode(member.member_id + '==').decode(), after.monitored_members)
+        )
+        for member, decoded in zip(after.monitored_members, decoded_user_ids, strict=True):
             print(f'id: {member.member_id} -> {decoded}')
-        self.assertTrue(not any('@' in d for d in decoded_user_ids), "wrong format for member IDs")
+        self.assertTrue(not any('@' in d for d in decoded_user_ids), 'wrong format for member IDs')